184 - project service refactoring
