+++ src/main/java/com/febfes/fftmback/controller/ProjectController.java
@@ -10,13 +10,15 @@ import com.febfes.fftmback.dto.OneProjectDto;
 import com.febfes.fftmback.dto.PatchDto;
 import com.febfes.fftmback.dto.ProjectDto;
 import com.febfes.fftmback.mapper.ProjectMapper;
-import com.febfes.fftmback.service.ProjectService;
 import com.febfes.fftmback.service.TaskTypeService;
+import com.febfes.fftmback.service.project.ProjectManagementService;
+import com.febfes.fftmback.service.project.ProjectMemberService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,7 +33,9 @@ import static com.febfes.fftmback.util.SortUtils.getOrderFromParams;
 @Tag(name = "Project")
 public class ProjectController {
 
-    private final @NonNull ProjectService projectService;
+    @Qualifier("projectManagementServiceDecorator")
+    private final @NonNull ProjectManagementService projectManagementService;
+    private final @NonNull ProjectMemberService projectMemberService;
     private final @NonNull TaskTypeService taskTypeService;
     private final @NonNull RoleCheckerComponent roleCheckerComponent;
 
@@ -41,7 +45,7 @@ public class ProjectController {
             @SortParam @RequestParam(defaultValue = "-createDate") String[] sort,
             @AuthenticationPrincipal UserEntity user
     ) {
-        return projectService.getProjectsForUser(user.getId(), getOrderFromParams(sort));
+        return projectMemberService.getProjectsForUser(user.getId(), getOrderFromParams(sort));
     }
 
     @Operation(summary = "Create new project")
@@ -50,7 +54,7 @@ public class ProjectController {
             @AuthenticationPrincipal UserEntity user,
             @RequestBody @Valid ProjectDto projectDto
     ) {
-        ProjectEntity project = projectService.createProject(
+        ProjectEntity project = projectManagementService.createProject(
                 ProjectMapper.INSTANCE.projectDtoToProject(projectDto), user.getId()
         );
         return ProjectMapper.INSTANCE.projectToProjectDto(project);
@@ -60,7 +64,7 @@ public class ProjectController {
     @ApiGetOne(path = "{id}")
     @SuppressWarnings("MVCPathVariableInspection") // fake warn "Cannot resolve path variable 'id' in @RequestMapping"
     public OneProjectDto getProject(@AuthenticationPrincipal UserEntity user, @PathVariable Long id) {
-        return projectService.getProjectForUser(id, user.getId());
+        return projectMemberService.getProjectForUser(id, user.getId());
     }
 
     @Operation(summary = "Edit project by its id")
@@ -70,14 +74,14 @@ public class ProjectController {
             @RequestBody ProjectDto projectDto
     ) {
         roleCheckerComponent.checkIfHasRole(id, RoleName.MEMBER_PLUS);
-        return projectService.editProject(id, ProjectMapper.INSTANCE.projectDtoToProject(projectDto));
+        return projectManagementService.editProject(id, ProjectMapper.INSTANCE.projectDtoToProject(projectDto));
     }
 
     @Operation(summary = "Delete project by its id")
     @ApiDelete(path = "{id}")
     public void deleteProject(@PathVariable Long id) {
         roleCheckerComponent.checkIfHasRole(id, RoleName.OWNER);
-        projectService.deleteProject(id);
+        projectManagementService.deleteProject(id);
     }
 
     @Operation(summary = "Edit project partially")
@@ -87,7 +91,7 @@ public class ProjectController {
             @PathVariable Long id,
             @RequestBody List<PatchDto> patchDtoList
     ) {
-        projectService.editProjectPartially(id, user.getId(), patchDtoList);
+        projectManagementService.editProjectPartially(id, user.getId(), patchDtoList);
     }
 
     @Operation(summary = "Get task types for project")
