+++ src/main/java/com/febfes/fftmback/service/project/DashboardService.java
@@ -1,0 +1,9 @@
+package com.febfes.fftmback.service.project;
+
+import com.febfes.fftmback.domain.common.specification.TaskSpec;
+import com.febfes.fftmback.dto.DashboardDto;
+
+public interface DashboardService {
+
+    DashboardDto getDashboard(Long id, TaskSpec taskSpec);
+}
