+++ src/main/java/com/febfes/fftmback/service/project/ProjectFavoriteService.java
@@ -1,0 +1,8 @@
+package com.febfes.fftmback.service.project;
+
+public interface ProjectFavoriteService {
+
+    void addProjectToFavourite(Long projectId, Long userId);
+
+    void removeProjectFromFavourite(Long projectId, Long userId);
+}
